+++ src/app/composables/dateTime.ts
@@ -2,7 +2,7 @@ import type { Dayjs } from 'dayjs'
 
 export const useDateTime = () => {
   const { $dayjs } = useNuxtApp()
-  const timezone = getTimezone()
+  const timezone = useTimezone()
 
   return (dateTime?: string | number | Dayjs | Date | null) =>
     $dayjs(dateTime).tz(timezone)