+++ src/server/utils/tusd.ts
@@ -53,3 +53,9 @@ export const deleteUpload = async (event: H3Event, uploadId: unknown) => {
   event.node.res.statusCode = 204
   await send(event)
 }
+
+export const useTusdFilesUrl = () => {
+  const { siteUrlTyped: siteUrl } = useSiteUrl()
+
+  return getTusdFilesUrl({ siteUrl })
+}