+++ src/shared/utils/tusd.ts
@@ -1,0 +1,2 @@
+export const getTusdFilesUrl = ({ siteUrl }: { siteUrl: URL }) =>
+  `https://tusd.${getDomainTldPort(siteUrl.host)}/files/`